Create script
A simple Google Apps Script used to call Google's AI API with prompt. Usage:

gCgAi(prompt,null);

Where prompt is the question text you want to ask AI. For this to work you need a google AI account and sign up for Google Apps Scripting, both of which are free. Here is how to get started:

https://www.google.com/script/start/
https://ai.google.dev/tutorials/ai-studio_quickstart